cgan/requirements.txt
asgiref==3.8.1
Django==4.2.20
filelock==3.18.0
fsspec==2025.3.0
image==1.5.33
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
pillow==11.1.0
six==1.17.0
sqlparse==0.5.3
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.12.2
